Refactor dashboard routing and component location
- Move UserDashboardPage from components to pages directory
- Update routing in App.tsx to nest dashboard route
- Add conditional rendering of UserDashboardPage in HNLiveTerminal
- Modify dashboard navigation to use 'navigate(-1)' for better UX
