+++ src/App.tsx
@@ -138,8 +138,8 @@ export function App() {
                 <Route path="terms" element={null} />
                 <Route path="privacy" element={null} />
                 <Route path="best-comments" element={null} />
+                <Route path="dashboard" element={null} />
               </Route>
-              <Route path="/dashboard" element={<UserDashboardPage />} />
             </Routes>
           </RunningStatusProvider>
         </AuthProvider>
